+++ test.clj
@@ -1,7 +1,29 @@
 ;;; font locking
 (ns clojure-mode.demo
-  (:require [clojure.something]
-            [something.s]))
+  (:require
+   [clojure.something]
+   [something.s]
+   [oneword]
+   ;; TODO mixedCase shouldn't be font-locked to clojure-interop-method-face
+   [mixedCase]
+   [CamelCase]))
+
+;; examples of valid namespace definitions
+(comment
+  (ns .validns)
+  (ns =validns)
+  (ns .ValidNs=<>?+|?*.)
+  (ns ValidNs<>?+|?*.b*ar.ba*z)
+  (ns other.valid.ns)
+  (ns oneword)
+  (ns one.X)
+  (ns foo.bar)
+  (ns Foo.bar)
+  (ns Foo.Bar)
+  (ns foo.Bar)
+  (ns Foo-bar)
+  (ns Foo-Bar)
+  (ns foo-Bar))
 
 (comment ;; for indentation
   (with-hi heya